// packages/ui/src/components/general/stat-widget/StatWidget.vue.d.ts
import Vue from 'vue';
/**
 * @displayName Stat Widget
 **/
export declare const ForgeStatWidget: import("vue/types/vue").ExtendedVue<Vue<Record<string, any>, Record<string, any>, never, never, (event: string, ...args: any[]) => Vue<Record<string, any>, Record<string, any>, never, never, any>>, unknown, unknown, {
    classes: string;
}, {
    size: string;
    variant: any;
}, {}, import("vue/types/v3-component-options").ComponentOptionsMixin, import("vue/types/v3-component-options").ComponentOptionsMixin>;
export default ForgeStatWidget;
//# sourceMappingURL=StatWidget.vue.d.ts.map